Added access check to DocumentRestController

+++ doclib-web/src/main/java/com/strelnikov/doclib/web/controllers/DocumentRestController.java
@@ -1,7 +1,10 @@
 package com.strelnikov.doclib.web.controllers;
 
 import com.strelnikov.doclib.dto.DocumentDto;
+import com.strelnikov.doclib.model.roles.PermissionType;
+import com.strelnikov.doclib.service.CatalogActions;
 import com.strelnikov.doclib.service.DocumentActions;
+import com.strelnikov.doclib.service.SecurityActions;
 import com.strelnikov.doclib.service.exceptions.UnitIsAlreadyExistException;
 import com.strelnikov.doclib.service.exceptions.UnitNotFoundException;
 import com.strelnikov.doclib.service.exceptions.VersionIsAlreadyExistException;
@@ -9,6 +12,8 @@ import com.strelnikov.doclib.service.exceptions.VersionNotExistException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.FileNotFoundException;
@@ -21,36 +26,56 @@ public class DocumentRestController {
     @Autowired
     DocumentActions docAct;
 
+    @Autowired
+    SecurityActions securityActions;
+
+    @Autowired
+    CatalogActions catalogActions;
+
     @GetMapping("get/{id}")
     public ResponseEntity<DocumentDto> getDocument(@PathVariable int id){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         try{
             DocumentDto docDto = docAct.loadDocument(id);
-            return ResponseEntity.ok(docDto);
+            if (securityActions.checkPermission(docDto,auth.getName(),PermissionType.WRITING)) {
+                return ResponseEntity.ok(docDto);
+            }
+            return ResponseEntity.status(403).build();
         } catch (UnitNotFoundException e){
             return ResponseEntity.notFound().build();
         }
     }
 
     @PostMapping("post")
-    public ResponseEntity<Object> postDocument(@RequestBody DocumentDto documentDto) throws VersionNotExistException {
+    public ResponseEntity<DocumentDto> postDocument(@RequestBody DocumentDto documentDto) throws VersionNotExistException {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         try {
-            documentDto = docAct.saveDocument(documentDto);
-            return ResponseEntity.ok(documentDto);
+            if(securityActions.checkPermission(catalogActions.loadCatalog(documentDto.getCatalogId()),auth.getName(), PermissionType.WRITING)){
+                documentDto = docAct.saveDocument(documentDto);
+                return ResponseEntity.ok(documentDto);
+            }
+            return ResponseEntity.status(403).build();
         } catch (VersionIsAlreadyExistException e) {
-            return ResponseEntity.badRequest().body("Version " + e.getDocVer().getVersion()+ " of document is already exist");
+            return ResponseEntity.badRequest().build();
         } catch (UnitIsAlreadyExistException e) {
-            return ResponseEntity.badRequest().body("Document is already exist in catalog");
+            return ResponseEntity.badRequest().build();
         } catch (FileNotFoundException e) {
-            return ResponseEntity.badRequest().body("One of the files doesn't exist");
+            return ResponseEntity.badRequest().build();
+        } catch (UnitNotFoundException e) {
+            return ResponseEntity.notFound().build();
         }
     }
 
     @DeleteMapping("delete/{id}")
-    public ResponseEntity<Object> deleteDocument(@PathVariable int id){
+    public ResponseEntity deleteDocument(@PathVariable int id){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         try {
-            docAct.loadDocument(id);
-            docAct.deleteDocument(id);
-            return ResponseEntity.ok("Document was successfully deleted");
+            DocumentDto documentDto = docAct.loadDocument(id);
+            if (securityActions.checkPermission(documentDto,auth.getName(),PermissionType.WRITING)) {
+                docAct.deleteDocument(id);
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(403).build();
         } catch (UnitNotFoundException e) {
             return ResponseEntity.notFound().build();
         }
